Move virtualfile-related tests into a separate test file
